requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python==4.8.1.78
requests==2.31.0
Pillow==10.1.0
pydantic==2.5.0

